backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
httpx==0.27.0
openai==1.37.0
python-jose==3.3.0
cryptography==43.0.1
